refactor: add different directories to group scripts
