+++ public/index.html
@@ -358,7 +358,7 @@
     </div>
     <br />
     <br />
-    <div class="ui footer huge menu">
+    <div id="footer" class="ui footer huge menu">
       <div class="left footer item">
         <div class="content">
           <div class="footer-title">
