+++ src/helpers/icalToJSON.js
