+++ src/models/blank.txt
@@ -1,0 +1,1 @@
+Temporary file
